refactor(acrawl-cli): rename render.rs to markdown.rs, relocate tool_format.rs
