+++ crates/acrawl-cli/src/tool_format.rs
