+++ crates/acrawl-cli/src/tui/repl_app.rs
@@ -9,7 +9,7 @@ use std::sync::{Arc, Mutex};
 use std::thread;
 use std::time::{Duration, Instant};
 
-use crate::render::PredictiveMarkdownBuffer;
+use crate::markdown::PredictiveMarkdownBuffer;
 use commands::{slash_command_specs, SlashCommand};
 use crossterm::cursor::SetCursorStyle;
 use crossterm::event::{
